Use Float for USD/EUR values

+++ src/main/scala/at/ac/ait/Model.scala
@@ -58,7 +58,7 @@ case class TxSummary(
 
 case class TxIdTime(height: Int, txHash: Array[Byte], timestamp: Int)
 
-case class Currency(satoshi: Long, eur: Double, usd: Double)
+case class Currency(satoshi: Long, eur: Float, usd: Float)
 
 case class AddressSummary(totalReceived: Currency, totalSpent: Currency)
